linter: move Report type and util funcs
- Move Report type to report.go

- Move utility functions from block.go and root.go to utils.go,
  since their usage (and scope) is not limited to either
  (and some of them are even exported).
